+++ internal/provider/model_iosxe_interface_ethernet.go
@@ -34,6 +34,10 @@ type InterfaceEthernet struct {
 	ChannelGroupNumber         types.Int64                        `tfsdk:"channel_group_number"`
 	ChannelGroupMode           types.String                       `tfsdk:"channel_group_mode"`
 	IpDhcpRelaySourceInterface types.String                       `tfsdk:"ip_dhcp_relay_source_interface"`
+	IpAccessGroupIn            types.String                       `tfsdk:"ip_access_group_in"`
+	IpAccessGroupInEnable      types.Bool                         `tfsdk:"ip_access_group_in_enable"`
+	IpAccessGroupOut           types.String                       `tfsdk:"ip_access_group_out"`
+	IpAccessGroupOutEnable     types.Bool                         `tfsdk:"ip_access_group_out_enable"`
 	HelperAddresses            []InterfaceEthernetHelperAddresses `tfsdk:"helper_addresses"`
 	SourceTemplate             []InterfaceEthernetSourceTemplate  `tfsdk:"source_template"`
 }
@@ -104,6 +108,22 @@ func (data InterfaceEthernet) toBody(ctx context.Context) string {
 	}
 	if !data.IpDhcpRelaySourceInterface.Null && !data.IpDhcpRelaySourceInterface.Unknown {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.dhcp.Cisco-IOS-XE-dhcp:relay.source-interface", data.IpDhcpRelaySourceInterface.Value)
+	}
+	if !data.IpAccessGroupIn.Null && !data.IpAccessGroupIn.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.in.acl.acl-name", data.IpAccessGroupIn.Value)
+	}
+	if !data.IpAccessGroupInEnable.Null && !data.IpAccessGroupInEnable.Unknown {
+		if data.IpAccessGroupInEnable.Value {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.in.acl.in", map[string]string{})
+		}
+	}
+	if !data.IpAccessGroupOut.Null && !data.IpAccessGroupOut.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.out.acl.acl-name", data.IpAccessGroupOut.Value)
+	}
+	if !data.IpAccessGroupOutEnable.Null && !data.IpAccessGroupOutEnable.Unknown {
+		if data.IpAccessGroupOutEnable.Value {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.access-group.out.acl.out", map[string]string{})
+		}
 	}
 	if len(data.HelperAddresses) > 0 {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address", []interface{}{})
@@ -206,6 +226,26 @@ func (data *InterfaceEthernet) updateFromBody(ctx context.Context, res gjson.Res
 		data.IpDhcpRelaySourceInterface.Value = value.String()
 	} else {
 		data.IpDhcpRelaySourceInterface.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupIn.Value = value.String()
+	} else {
+		data.IpAccessGroupIn.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.in"); value.Exists() {
+		data.IpAccessGroupInEnable.Value = true
+	} else {
+		data.IpAccessGroupInEnable.Value = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupOut.Value = value.String()
+	} else {
+		data.IpAccessGroupOut.Null = true
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.out"); value.Exists() {
+		data.IpAccessGroupOutEnable.Value = true
+	} else {
+		data.IpAccessGroupOutEnable.Value = false
 	}
 	for i := range data.HelperAddresses {
 		keys := [...]string{"address"}
@@ -341,6 +381,28 @@ func (data *InterfaceEthernet) fromBody(ctx context.Context, res gjson.Result) {
 		data.IpDhcpRelaySourceInterface.Value = value.String()
 		data.IpDhcpRelaySourceInterface.Null = false
 	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupIn.Value = value.String()
+		data.IpAccessGroupIn.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.in.acl.in"); value.Exists() {
+		data.IpAccessGroupInEnable.Value = true
+		data.IpAccessGroupInEnable.Null = false
+	} else {
+		data.IpAccessGroupInEnable.Value = false
+		data.IpAccessGroupInEnable.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.acl-name"); value.Exists() {
+		data.IpAccessGroupOut.Value = value.String()
+		data.IpAccessGroupOut.Null = false
+	}
+	if value := res.Get(prefix + "ip.access-group.out.acl.out"); value.Exists() {
+		data.IpAccessGroupOutEnable.Value = true
+		data.IpAccessGroupOutEnable.Null = false
+	} else {
+		data.IpAccessGroupOutEnable.Value = false
+		data.IpAccessGroupOutEnable.Null = false
+	}
 	if value := res.Get(prefix + "ip.helper-address"); value.Exists() {
 		data.HelperAddresses = make([]InterfaceEthernetHelperAddresses, 0)
 		value.ForEach(func(k, v gjson.Result) bool {
@@ -444,6 +506,22 @@ func (data *InterfaceEthernet) setUnknownValues(ctx context.Context) {
 		data.IpDhcpRelaySourceInterface.Unknown = false
 		data.IpDhcpRelaySourceInterface.Null = true
 	}
+	if data.IpAccessGroupIn.Unknown {
+		data.IpAccessGroupIn.Unknown = false
+		data.IpAccessGroupIn.Null = true
+	}
+	if data.IpAccessGroupInEnable.Unknown {
+		data.IpAccessGroupInEnable.Unknown = false
+		data.IpAccessGroupInEnable.Null = true
+	}
+	if data.IpAccessGroupOut.Unknown {
+		data.IpAccessGroupOut.Unknown = false
+		data.IpAccessGroupOut.Null = true
+	}
+	if data.IpAccessGroupOutEnable.Unknown {
+		data.IpAccessGroupOutEnable.Unknown = false
+		data.IpAccessGroupOutEnable.Null = true
+	}
 	for i := range data.HelperAddresses {
 		if data.HelperAddresses[i].Address.Unknown {
 			data.HelperAddresses[i].Address.Unknown = false
@@ -529,6 +607,12 @@ func (data *InterfaceEthernet) getEmptyLeafsDelete(ctx context.Context) []string
 	emptyLeafsDelete := make([]string, 0)
 	if !data.Shutdown.Value {
 		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/shutdown", data.getPath()))
+	}
+	if !data.IpAccessGroupInEnable.Value {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/ip/access-group/in/apply-type/apply-intf/acl/in", data.getPath()))
+	}
+	if !data.IpAccessGroupOutEnable.Value {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/ip/access-group/out/apply-type/apply-intf/acl/out", data.getPath()))
 	}
 
 	for i := range data.HelperAddresses {