+++ internal/provider/resource_iosxe_interface_ethernet_test.go
@@ -22,6 +22,10 @@ func TestAccIosxeInterfaceEthernet(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ipv4_address", "15.1.1.1"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ipv4_address_mask", "255.255.255.252"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_dhcp_relay_source_interface", "Loopback100"),
+					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_access_group_in", "1"),
+					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_access_group_in_enable", "true"),
+					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_access_group_out", "1"),
+					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_access_group_out_enable", "true"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.address", "10.10.10.10"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.global", "false"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.vrf", "VRF1"),
@@ -78,6 +82,10 @@ func testAccIosxeInterfaceEthernetConfig_all() string {
 		ipv4_address = "15.1.1.1"
 		ipv4_address_mask = "255.255.255.252"
 		ip_dhcp_relay_source_interface = "Loopback100"
+		ip_access_group_in = "1"
+		ip_access_group_in_enable = true
+		ip_access_group_out = "1"
+		ip_access_group_out_enable = true
 		helper_addresses = [{
 		address = "10.10.10.10"
 		global = false