+++ docs/guides/changelog.md
@@ -11,6 +11,7 @@ description: |-
 
 - Add iosxe_access_list_standard resource and data source
 - Add iosxe_access_list_extended resource and data source
+- Add ip_access_group_* attributes to iosxe_interface_ethernet resource and data source
 
 ## 0.1.10